+++ frontends/p4/metrics/matchActionTableMetrics.cpp
@@ -47,11 +47,11 @@ void MatchActionTableMetricsPass::postorder(const IR::P4Table *table) {
 
 void MatchActionTableMetricsPass::postorder(const IR::P4Program* /*program*/){
     if(metrics.matchActionTableMetrics.numTables > 0){
-        metrics.matchActionTableMetrics.avgKeysPerTable = static_cast<double>(metrics.matchActionTableMetrics.totalKeys) / metrics.matchActionTableMetrics.numTables;
-        metrics.matchActionTableMetrics.avgActionsPerTable = static_cast<double>(metrics.matchActionTableMetrics.totalActions) / metrics.matchActionTableMetrics.numTables;
+        metrics.matchActionTableMetrics.avgKeysPerTable = static_cast<double>(metrics.matchActionTableMetrics.totalKeys) / static_cast<double>(metrics.matchActionTableMetrics.numTables);
+        metrics.matchActionTableMetrics.avgActionsPerTable = static_cast<double>(metrics.matchActionTableMetrics.totalActions) / static_cast<double>(metrics.matchActionTableMetrics.numTables);
     }
     if (metrics.matchActionTableMetrics.totalKeys > 0)
-        metrics.matchActionTableMetrics.avgKeySize = static_cast<double>(metrics.matchActionTableMetrics.totalKeySizeSum) / metrics.matchActionTableMetrics.totalKeys;
+        metrics.matchActionTableMetrics.avgKeySize = static_cast<double>(metrics.matchActionTableMetrics.totalKeySizeSum) / static_cast<double>(metrics.matchActionTableMetrics.totalKeys);
 }
 
 }  // namespace P4
